flexutils/requirements/env_requirements.txt
scipion-pyworkflow
scipion-em >= 3.0.8
numpy==1.18.4
scipion-em-xmipp
pyvista==0.27.4
pyvistaqt==0.3.0
pymeshfix==0.14.1
scikit-image==0.18.1
PyQt5==5.15.4
mayavi==4.7.4
shapely==1.8.1.post1
umap-learn==0.5.1
mlrose==1.3.0
prody==2.2.0